+++ packages/server/src/supertokens.test.ts
@@ -15,7 +15,7 @@ import {
   TOKEN_SEPARATOR,
   SessionContext,
   parsePublicDataToken,
-} from "."
+} from "./supertokens"
 
 const isIsoDate = (str: string) => {
   if (!/\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}.\d{3}Z/.test(str)) return false
