--- packages/supertokens/README.md
@@ -1,1 +1,0 @@
-# @blitzjs/supertokens
