--- packages/supertokens/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "@blitzjs/supertokens",
-  "description": "The Blitz.js session management library",
-  "version": "0.15.0",
-  "author": "Brandon Bayer <b@bayer.ws>",
-  "license": "MIT",
-  "scripts": {
-    "clean": "rimraf dist",
-    "dev": "tsdx watch --verbose",
-    "build": "tsdx build",
-    "test": "tsdx test"
-  },
-  "main": "dist/index.js",
-  "module": "dist/supertokens.esm.js",
-  "typings": "dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/blitz-js/blitz"
-  },
-  "dependencies": {
-    "hasha": "5.2.0",
-    "cookie": "0.4.1",
-    "date-fns": "2.14.0",
-    "tiny-invariant": "1.1.0"
-  },
-  "gitHead": "d3b9fce0bdd251c2b1890793b0aa1cd77c1c0922"
-}
